pyproject.toml
[tool.poetry]
name = "auto-labeler"
version = "0.1.0"
description = "A simple tool for automatically labeling and archiving emails based on your past activity"
authors = ["Brian Smith <ohohvi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
google-api-python-client = "^2.104.0"
google-auth-httplib2 = "^0.1.1"
google-auth-oauthlib = "^1.1.0"
appdirs = "^1.4.4"
tabulate = "^0.9.0"
beautifulsoup4 = "^4.12.2"
model-personal-trainer = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
auto-labeler = "auto_labeler.main:main"
